Fix code bugs
