+++ README.md
@@ -120,8 +120,8 @@ func main() {
 	// Start the engine.
 	engine.Run()
 
-	// Wait for 10 seconds. you can use `curl http://127.0.0.1:8080/ping` to test.
-	time.Sleep(10 * time.Second)
+	// Wait for 30 seconds.
+	time.Sleep(30 * time.Second)
 
 	// Stop the engine.
 	engine.Stop()
@@ -155,18 +155,75 @@ successs
 
 Because `orbit` is based on `gin`, so you can use `gin` middleware directly. So you can use custom middleware to do some custom things. For example, you can use `cors` middleware to support `cors` request.
 
-Here is a example to use `demo` middleware to print `demo` in the console.
+Here is a example to use `demo` middleware to print `>>>>>>!!! demo` in the console.
 
 **Example**
 
 ```go
+package main
 
+import (
+	"fmt"
+	"time"
+
+	"github.com/gin-gonic/gin"
+	"github.com/shengyanli1982/orbit"
+)
+
+func customMiddleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		fmt.Println(">>>>>>!!! demo")
+		c.Next()
+	}
+}
+
+type service struct{}
+
+func (s *service) RegisterGroup(g *gin.RouterGroup) {
+	g.GET("/demo", func(c *gin.Context) {})
+}
+
+func main() {
+	// Create a new orbit configuration.
+	config := orbit.NewConfig()
+
+	// Create a new orbit feature options.
+	opts := orbit.NewOptions().EnableMetric()
+
+	// Create a new orbit engine.
+	engine := orbit.NewEngine(config, opts)
+
+	// Register a custom middleware.
+	engine.RegisterMiddleware(customMiddleware())
+
+	// Register a custom router group.
+	engine.RegisterService(&service{})
+
+	// Start the engine.
+	engine.Run()
+
+	// Wait for 30 seconds.
+	time.Sleep(30 * time.Second)
+
+	// Stop the engine.
+	engine.Stop()
+}
 ```
 
 **Result**
 
 ```bash
+$ go run demo.go
+[GIN-debug] [WARNING] Running in "debug" mode. Switch to "release" mode in production.
+ - using env:	export GIN_MODE=release
+ - using code:	gin.SetMode(gin.ReleaseMode)
 
+[GIN-debug] GET    /ping                     --> github.com/shengyanli1982/orbit.healthcheckService.func1 (1 handlers)
+[GIN-debug] GET    /metrics                  --> github.com/shengyanli1982/orbit/utils/wrapper.WrapHandlerToGin.func1 (2 handlers)
+[GIN-debug] GET    /demo                     --> main.(*service).RegisterGroup.func1 (7 handlers)
+{"level":"INFO","time":"2024-01-10T20:03:38.869+0800","logger":"default","caller":"orbit/gin.go:162","message":"http server is ready","address":"127.0.0.1:8080"}
+>>>>>>!!! demo
+{"level":"INFO","time":"2024-01-10T20:03:41.275+0800","logger":"default","caller":"log/default.go:10","message":"http server access log","id":"","ip":"127.0.0.1","endpoint":"127.0.0.1:59787","path":"/demo","method":"GET","code":200,"status":"OK","latency":"780ns","agent":"curl/8.1.2","query":"","reqContentType":"","reqBody":""}
 ```
 
 ## 5. Custom Router Group
